% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{estimate_quantile.minmi}
\alias{estimate_quantile.minmi}
\title{Estimate a specific quantile using the MINMI method.}
\usage{
estimate_quantile.minmi(K, W, u = NULL, eps.sigma = 0, q)
}
\arguments{
\item{K}{Numeric upper bound for fossil ages.}

\item{W}{Numeric vector of fossil ages.}

\item{u}{Optional Matrix of numbers sampled from a uniform distribution from 0 to 1. Rows correspond to fossils and Columns correspond to Monte Carlo samples}

\item{eps.sigma}{Optional vector of standard errors for each fossil. Defaults to 0.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}
}
\value{
A numeric indicating the MINMI estimate of the qth quantile of our sample statistic.
}
\description{
Estimate a specific quantile using the MINMI method.
}
