% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{estimating_eqn}
\alias{estimating_eqn}
\title{Estimating equation for Newton-Raphson's Algorithm}
\usage{
estimating_eqn(theta, q, K, u, m, n, eps.mean, eps.sigma)
}
\arguments{
\item{theta}{Number. Estimate for theta.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}

\item{K}{Numeric. Upper bound for fossil record.}

\item{u}{Numeric vector of values sampled from a uniform distribution from 0 to 1 and with same length as W.}

\item{m}{Numeric. Sample minimum from fossil record.}

\item{n}{Number of fossils.}

\item{eps.mean}{Numeric mean of the measurement error distribution. Defaults to 0.}

\item{eps.sigma}{Numeric standard error of the measurement error distribution. Defaults to 0.}
}
\value{
A number
}
\description{
Estimating equation for Newton-Raphson's Algorithm
}
