% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{find_optimal_B}
\alias{find_optimal_B}
\title{Function for Finding the Number of Monte Carlo Samples for MINMI}
\usage{
find_optimal_B(max_var, K, m, n, u, eps.sigma, q)
}
\arguments{
\item{max_var}{Numeric greater than 0 specifying the maximum Monte Carlo error variance we aim to have associated with our MINMI estimates.}

\item{K}{Numeric. Upper bound for fossil record.}

\item{m}{Numeric. Sample minimum from fossil record.}

\item{n}{Number of fossils.}

\item{u}{Matrix of numbers sampled from a uniform distribution from 0 to 1. Rows correspond to fossils and Columns correspond to Monte Carlo samples}

\item{eps.sigma}{Vector of standard errors for each fossil. Length n.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}
}
\value{
Vector of numbers for each fossil.
}
\description{
Function for Finding the Number of Monte Carlo Samples for MINMI
}
