% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{minmi}
\alias{minmi}
\title{Estimate Extinction Times using the MINMI Estimator}
\usage{
minmi(ages, sd, K, alpha = 0.05, B = NULL, A = NULL, .B_init = 500)
}
\arguments{
\item{ages}{Numeric vector of fossil ages.}

\item{sd}{Numeric vector of measurement error uncertainty associated with each fossil.}

\item{K}{Numeric upper bound for fossil ages.}

\item{alpha}{Numeric between 0 and 1. Used to find 100(1-alpha)\% confidence intervals. Defaults to 0.05 (95\% confidence intervals)}

\item{B}{Optional numeric greater than 1 specifying the number of Monte Carlo samples to use.}

\item{A}{Optional numeric greater than 0 specifying the maximum Monte Carlo error variance we aim to have associated with our MINMI estimates.}

\item{.B_init}{Optional numeric greater than 1 specifying the number of Monte Carlo samples to use in the pilot estimates. Defaults to 500.}
}
\value{
minmi() returns a list with estimates for the lower end point of the 100(1-alpha)\% confidence interval, point estimate, upper end point, and a list containing the B's used for each.
}
\description{
A generic function for estimating extinction times using the MINMI procedure
}
\details{
Note that providing a value for `B` means that the value for `A` is made redundant, as `A` is used to find `B` but not the other way around.
}
\examples{
ages = runif(20, 10000, 25000)
sd = runif(20, 50, 100)

minmi(ages=ages, sd=sd, K=22000, alpha=0.05)
minmi(ages=ages, sd=sd, K=22000, alpha=0.05, B = 100)
minmi(ages=ages, sd=sd, K=22000, alpha=0.05, A = 1000)
}
