% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{uniform_to_tnorm}
\alias{uniform_to_tnorm}
\title{Helper function to transform samples from a uniform dist to a truncated normal}
\usage{
uniform_to_tnorm(u, eps.sigma, a, b)
}
\arguments{
\item{u}{Matrix of numbers. Rows = fossils; columns = MC samples}

\item{eps.sigma}{Vector of numbers. Standard deviations for each fossil.}

\item{a}{Lower bound for truncated normal}

\item{b}{Upper bound for truncated normal}
}
\value{
Matrix of numbers. Same dimensions as u
}
\description{
Helper function to transform samples from a uniform dist to a truncated normal
}
