% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_B.R
\name{choose_B}
\alias{choose_B}
\title{Function for Finding the Number of Monte Carlo Samples for MINMI}
\usage{
choose_B(A, K, m, n, u, eps.sigma, q)
}
\arguments{
\item{A}{Numeric greater than 0 specifying the maximum Monte Carlo error variance we aim to have associated with our MINMI estimates.}

\item{K}{Numeric upper bound for fossil ages.}

\item{m}{Numeric. Sample minimum from fossil record.}

\item{n}{Number of fossils.}

\item{u}{Matrix of numbers sampled from a uniform distribution from 0 to 1. Rows correspond to fossils and Columns correspond to Monte Carlo samples}

\item{eps.sigma}{Vector of standard errors for each fossil. Length n.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}
}
\value{
An integer representing the number of Monte Carlo samples to use.
}
\description{
Function for Finding the Number of Monte Carlo Samples for MINMI
}
