% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{estimating_eqn}
\alias{estimating_eqn}
\title{Estimating equation for Newton-Raphson's Algorithm}
\usage{
estimating_eqn(theta, q, K, u, m, n, eps.sigma)
}
\arguments{
\item{theta}{Number. Estimate for theta.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}

\item{K}{Numeric. Upper bound for fossil record.}

\item{u}{Matrix of numbers sampled from a uniform distribution from 0 to 1. Rows correspond to fossils and Columns correspond to Monte Carlo samples}

\item{m}{Numeric. Sample minimum from fossil record.}

\item{n}{Number of fossils.}

\item{eps.sigma}{Vector of standard errors for each fossil. Length n.}
}
\value{
A number
}
\description{
Estimating equation for Newton-Raphson's Algorithm
}
