% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{find_optimal_B}
\alias{find_optimal_B}
\title{Function for Finding the Number of Monte Carlo Samples for MINMI}
\usage{
find_optimal_B(max_var, K, m, n, u, eps.mean, eps.sigma, q)
}
\arguments{
\item{max_var}{Numeric greater than 0 specifying the maximum Monte Carlo error variance we aim to have associated with our MINMI estimates.}

\item{K}{Numeric. Upper bound for fossil record.}

\item{m}{Numeric. Sample minimum from fossil record.}

\item{n}{Number of fossils.}

\item{u}{Numeric vector of values sampled from a uniform distribution from 0 to 1 and with same length as W.}

\item{eps.mean}{Numeric mean of the measurement error distribution. Defaults to 0.}

\item{eps.sigma}{Numeric standard error of the measurement error distribution. Defaults to 0.}

\item{q}{Numeric between 0 and 1 specifying the quantile we want to estimate.}
}
\value{
A number.
}
\description{
Function for Finding the Number of Monte Carlo Samples for MINMI
}
