% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmi.R
\name{minmi}
\alias{minmi}
\title{Estimate Extinction Times using the MINMI Estimator}
\usage{
minmi(W, sd, K, alpha = 0.05, B = NULL, .B_init = 500, .max_var = NULL)
}
\arguments{
\item{W}{Numeric vector of fossil ages.}

\item{sd}{Numeric vector of measurement error uncertainty associated with each fossil.}

\item{K}{Numeric upper bound for fossil ages.}

\item{alpha}{Numeric between 0 and 1. Used to find 100(1-alpha)% confidence intervals. Defaults to 0.05 (95% confidence intervals)}

\item{B}{Optional numeric greater than 1 specifying the number of Monte Carlo samples to use.}

\item{.B_init}{Optional numeric greater than 1 specifying the number of Monte Carlo samples to use in the pilot estimates. Defaults to 500.}

\item{.max_var}{Optional numeric greater than 0 specifying the maximum Monte Carlo error variance we aim to have associated with our MINMI estimates.}
}
\value{
A list with estimates for the lower end point of the 100(1-alpha)% confidence interval, point estimate, upper end point, and a list containing the B's used for each.
}
\description{
This function estimates extinction times using the MINMI procedure
}
